# openapi.yaml
openapi: 3.0.3
info:
  title: FinApp API
  description: API for financial application
  version: 1.3.0
  contact:
    name: API Support
    email: support@finapp.com
servers:
  - url: http://localhost:8080/api/v1
    description: Local development server
  - url: https://api.finapp.com/api/v1
    description: Production server
paths:
...
  /receipts:
    get:
      operationId: getReceipts
      tags: [Receipts]
      summary: Get receipts list
      description: |
        Получает пагинированный список чеков пользователя.
      parameters:
        - name: page
          in: query
          description: Page number (0-based)
          required: false
          schema:
            type: integer
            minimum: 0
            default: 0
        - name: size
          in: query
          description: Page size
          required: false
          schema:
            type: integer
            minimum: 1
            maximum: 100
            default: 20
        - name: transactionId
          in: query
          description: Filter by transaction ID
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Receipts list
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PagedReceiptResponse'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    post:
      operationId: uploadReceipt
      tags: [Receipts]
      summary: Upload receipt
      description: |
        Загружает файл чека (изображение или PDF) с возможностью привязки к транзакции.
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                file:
                  type: string
                  format: binary
                  description: Файл чека (изображение или PDF)
                transactionId:
                  type: string
                  format: uuid
                  description: ID транзакции для привязки
                description:
                  type: string
                  description: Описание чека
              required: [file]
      responses:
        '201':
          description: Receipt uploaded successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/UploadReceiptResponse'
        '400':
          description: Invalid file format
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /receipts/{id}:
    get:
      operationId: getReceiptById
      tags: [Receipts]
      summary: Get receipt by ID
      description: |
        Получает информацию о чеке по его идентификатору.
      parameters:
        - name: id
          in: path
          description: Receipt ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Receipt details
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptResponse'
        '404':
          description: Receipt not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    put:
      operationId: updateReceipt
      tags: [Receipts]
      summary: Update receipt
      description: |
        Обновляет информацию о чеке.
      parameters:
        - name: id
          in: path
          description: Receipt ID
          required: true
          schema:
            type: string
            format: uuid
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateReceiptRequest'
      responses:
        '200':
          description: Receipt updated successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ReceiptResponse'
        '404':
          description: Receipt not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
    delete:
      operationId: deleteReceipt
      tags: [Receipts]
      summary: Delete receipt
      description: |
        Удаляет чек и связанный с ним файл.
      parameters:
        - name: id
          in: path
          description: Receipt ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '204':
          description: Receipt deleted successfully
        '404':
          description: Receipt not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
  /receipts/{id}/download:
    get:
      operationId: downloadReceipt
      tags: [Receipts]
      summary: Download receipt file
      description: |
        Скачивает файл чека.
      parameters:
        - name: id
          in: path
          description: Receipt ID
          required: true
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: Receipt file
          content:
            application/octet-stream:
              schema:
                type: string
                format: binary
            image/*:
              schema:
                type: string
                format: binary
            application/pdf:
              schema:
                type: string
                format: binary
        '404':
          description: Receipt not found
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
        '401':
          description: Unauthorized
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'

  # Dictionary endpoints
...
components:
  schemas:
...
    ProblemDetails:
      type: object
      required: [type, title, status]
      properties:
        type:
          type: string
          format: uri
          description: "URI идентифицирующий тип проблемы"
          example: "https://api.finapp.com/problems/validation-error"
        title:
          type: string
          description: "Краткое описание типа проблемы"
          example: "Validation Error"
        status:
          type: integer
          description: "HTTP статус код"
          example: 400
        detail:
          type: string
          description: "Детальное описание проблемы"
          example: "Поле 'amount' должно быть положительным числом"
        instance:
          type: string
          format: uri
          description: "URI идентифицирующий конкретный случай проблемы"
          example: "/api/v1/transactions/123"
        errors:
          type: array
          description: "Детали ошибок валидации"
          items:
            type: object
            properties:
              field:
                type: string
                description: "Поле с ошибкой"
                example: "amount"
              message:
                type: string
                description: "Сообщение об ошибке"
                example: "должно быть положительным числом"
              code:
                type: string
                description: "Код ошибки"
                example: "POSITIVE_NUMBER_REQUIRED"

    # User DTOs
...
    ReceiptResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        originalFilename:
          type: string
        contentType:
          type: string
        fileSize:
          type: integer
        uploadedAt:
          type: string
          format: date-time
        transactionId:
          type: string
          format: uuid
          nullable: true
        description:
          type: string
          nullable: true
        downloadUrl:
          type: string
          description: "URL для скачивания"
    PagedReceiptResponse:
      type: object
      properties:
        content:
          type: array
          items:
            $ref: '#/components/schemas/ReceiptResponse'
        totalElements:
          type: integer
        totalPages:
          type: integer
        number:
          type: integer
        size:
          type: integer
        first:
          type: boolean
        last:
          type: boolean
    UploadReceiptResponse:
      type: object
      properties:
        id:
          type: string
          format: uuid
        filename:
          type: string
        contentType:
          type: string
        fileSize:
          type: integer
        uploadedAt:
          type: string
          format: date-time
        downloadUrl:
          type: string
    UpdateReceiptRequest:
      type: object
      properties:
        description:
          type: string
          description: "Описание чека"
        transactionId:
          type: string
          format: uuid
          nullable: true
          description: "ID транзакции для привязки"

    # Dictionary DTOs
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
security:
  - BearerAuth: []